login/serializer.py
from .models import Users
from rest_framework import serializers


class UsersSerializer(serializers.ModelSerializer):
    class Meta:
        model = Users
        fields = ("id", "username", "password", "name", "email", "created_on")