requirements.txt
asgiref==3.1.2
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
channels==2.2.0
constantly==15.1.0
daphne==2.3.0
Django==2.2.1
django-redis==4.10.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pkg-resources==0.0.0
psycopg2==2.8.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
zope.interface==4.6.0
